# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restau TyTy API",
        "description": "API documentation for Restaurant Management System",
        "contact": {
            "email": "admin@restautyty.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://restau-bo.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/commandes/{commandeId}/items": {
            "get": {
                "tags": [
                    "CommandeItems"
                ],
                "summary": "Get list of commande items",
                "operationId": "getCommandeItemsList",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "CommandeItems"
                ],
                "summary": "Add item to commande",
                "description": "Guest requires X-Order-Token",
                "operationId": "storeCommandeItem",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Order-Token",
                        "in": "header",
                        "description": "Guest Session Token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_type",
                                    "item_id",
                                    "quantite"
                                ],
                                "properties": {
                                    "item_type": {
                                        "type": "string",
                                        "enum": [
                                            "plat",
                                            "boisson"
                                        ]
                                    },
                                    "item_id": {
                                        "type": "integer"
                                    },
                                    "quantite": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/commandes/{commandeId}/items/{id}": {
            "get": {
                "tags": [
                    "CommandeItems"
                ],
                "summary": "Get commande item information",
                "operationId": "getCommandeItemById",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Order-Token",
                        "in": "header",
                        "description": "Guest Session Token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "CommandeItems"
                ],
                "summary": "Update commande item",
                "operationId": "updateCommandeItem",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommandeItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CommandeItems"
                ],
                "summary": "Delete commande item",
                "description": "Guest requires X-Order-Token",
                "operationId": "deleteCommandeItem",
                "parameters": [
                    {
                        "name": "commandeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-Order-Token",
                        "in": "header",
                        "description": "Guest Session Token",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommandeItem": {
                "required": [
                    "commande_id",
                    "itemable_id",
                    "itemable_type",
                    "quantite",
                    "prix_unitaire"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "commande_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "itemable_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "itemable_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\Plat"
                    },
                    "quantite": {
                        "type": "integer",
                        "example": 2
                    },
                    "prix_unitaire": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}